Change css to a module, fusion files
