+++ ui/src/components/ChatWidget.module.css
@@ -1,65 +1,84 @@
-.rcw-widget-container {
+.widgetContainer {
   position: fixed;
   right: 20px;
   bottom: 20px;
   width: 360px;
-  max-width: calc(100% - 40px);
+  max-width: 415px;
   background: #fff;
   border-radius: 8px;
   box-shadow: 0 10px 30px rgba(0, 0, 0, 0.12);
   display: flex;
   flex-direction: column;
   font-family: inherit;
-  z-index: 2000;
+  z-index: 100002;
+  cursor: move;
+  color: black;
+  padding: 1em;
+  margin: auto;
+  user-select: none;
 }
 
-.rcw-hidden {
+.hidden {
   display: none;
 }
 
-.rcw-header {
+.conversationContainer {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
+}
+
+.header {
   padding: 0.6rem;
-  border-bottom: 1px solid #eee;
-  background: #f7f7f8;
+  border-bottom: 1px solid #ddd;
+  background: #f4f7f9;
+  color: rgb(90, 85, 85);
   position: relative;
 }
-.rcw-title {
+
+.title {
   font-weight: 700;
   font-size: 1rem;
 }
-.rcw-subtitle {
+
+.subtitle {
   font-size: 0.85rem;
   color: #666;
 }
 
-.rcw-messages-container {
+.messagesContainer {
   padding: 0.6rem;
   max-height: 280px;
   overflow: auto;
-  background: #fafafa;
+  background: #f9f9f9;
 }
 
-.rcw-message {
+.message {
   margin-bottom: 0.6rem;
   display: flex;
   flex-direction: column;
 }
-.rcw-response {
+
+.response {
   align-items: flex-start;
 }
-.rcw-message-client {
+
+.messageClient {
   align-items: flex-end;
 }
-.rcw-client .rcw-message-text {
+
+.client .messageText {
   background: #cfeef3;
   color: #000;
 }
-.rcw-response .rcw-message-text {
+
+.response .messageText {
   background: #fff;
   color: #000;
 }
 
-.rcw-message-text {
+.messageText {
   padding: 0.45rem 0.6rem;
   border-radius: 6px;
   max-width: 85%;
@@ -67,49 +86,51 @@
   box-shadow: 0 1px 0 rgba(0, 0, 0, 0.03);
 }
 
-.rcw-timestamp {
+.timestamp {
   font-size: 0.72rem;
   color: #999;
   margin-top: 0.25rem;
 }
 
-.rcw-sender {
+.sender {
   padding: 0.6rem;
-  border-top: 1px solid #eee;
+  border-top: 1px solid #ddd;
   display: flex;
   gap: 0.5rem;
   align-items: flex-end;
 }
-.rcw-new-message {
+
+.newMessage {
   flex: 1;
   padding: 0.4rem;
   border-radius: 6px;
   border: 1px solid #e6e6e6;
   resize: none;
 }
-.rcw-send {
+
+.send {
   padding: 0.35rem 0.7rem;
 }
 
-.rcw-launcher {
+.launcher {
   position: fixed;
   right: 20px;
   bottom: 20px;
   cursor: pointer;
   z-index: 2001;
-  background: none;
+  background-color: transparent;
   border: none;
   padding: 0;
   outline: none;
   box-shadow: none;
 }
 
-.rcw-launcher:focus {
+.launcher:focus {
   outline: none;
   box-shadow: none;
 }
 
-.rcw-badge {
+.badge {
   background: red;
   color: #fff;
   min-width: 20px;
@@ -121,4 +142,16 @@
   font-size: 12px;
   font-weight: 700;
   padding: 0 6px;
+}
+
+.chatWidgetDragable {
+  position: absolute;
+  bottom: 5px;
+  right: 0;
+  z-index: 100002;
+}
+
+.chatWidgetDragable .reactDragable {
+  position: relative;
+  z-index: 100002;
 }
