--- ui/src/components/rachat.css
@@ -1,81 +1,0 @@
-.chat-widget-dragable {
-  position: absolute;
-  bottom: 5px;
-  right: 0;
-  z-index: 100002;
-}
-
-.chat-widget-dragable .react-dragable {
-  position: relative;
-  z-index: 100002;
-}
-
-.rcw-widget-container {
-  z-index: 10001;
-  max-width: 415px;
-  float: right;
-  cursor: move;
-  color: black;
-  border-radius: 5px;
-  padding: 1em;
-  margin: auto;
-  user-select: none;
-}
-
-.rcw-conversation-container .rcw-header {
-  background-color: #f4f7f9;
-  color: rgb(90, 85, 85);
-  padding: 0;
-  border-bottom: 1px solid #ddd;
-}
-
-.rcw-conversation-container > .rcw-sender {
-  border-top: 1px solid #ddd;
-}
-
-.rcw-messages-container {
-  background-color: #f9f9f9;
-}
-
-.rcw-messages-container > .rcw-message > .rcw-client {
-  margin-left: unset;
-  border: 1px solid #999;
-  border-radius: 4px;
-  box-shadow: 0 1px 1px rgba(0, 0, 0, 0.2);
-  background-color: #c6eab7;
-  padding: 10px;
-  width: 100%;
-}
-
-.rcw-messages-container > .rcw-message > .rcw-response {
-  margin-left: auto;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  box-shadow: 0 1px 1px rgba(0, 0, 0, 0.095);
-  background-color: #fff;
-  padding: 10px;
-  width: 100%;
-}
-
-.rcw-messages-container > .rcw-message > .rcw-response {
-  margin-left: auto;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  box-shadow: 0 1px 1px rgba(0, 0, 0, 0.095);
-  background-color: #fff;
-  padding: 10px;
-  width: 100%;
-}
-
-.rcw-widget-container .rcw-launcher {
-  background-color: #3766ad;
-  border: 1px solid #000;
-}
-
-.rcw-launcher .rcw-badge {
-  width: 22px;
-  height: 22px;
-  line-height: 14px;
-  background-color: #0366d6;
-  border: 1px solid #000;
-}
